+++ boot/core/src/boot/core.clj
@@ -536,7 +536,7 @@
         (let [q       (LinkedBlockingQueue.)
               watcher (apply file/watcher! :time dirs)
               paths   (into-array String dirs)
-              k       (.invoke pod/worker-pod "boot.watcher/make-watcher" q paths)]
+              k       (pod/with-invoke-worker (boot.watcher/make-watcher q paths))]
           (daemon
             (loop [ret (util/guard [(.take q)])]
               (when ret
@@ -545,7 +545,7 @@
                   (let [changed (watcher)]
                     (when-not (empty? changed) (callback changed))
                     (recur (util/guard [(.take q)])))))))
-          #(.invoke pod/worker-pod "boot.watcher/stop-watcher" k)))))
+          #(pod/with-invoke-worker (boot.watcher/stop-watcher k))))))
 
 (defn rebuild!
   "Manually trigger build watch."